pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "faq-bot"
version = "2.0.0"
description = "Intelligent Telegram FAQ Bot with Semantic Search"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "FAQ Bot Development Team", email = "support@faqbot.example.com"},
]
maintainers = [
    {name = "FAQ Bot Development Team", email = "support@faqbot.example.com"},
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Communications :: Chat",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "faiss-cpu>=1.7.0",
    "sentence-transformers>=2.2.0",
    "aiogram",
    "pandas",
    "openpyxl",
    "tkinterdnd2>=0.3.0",
    "python-dotenv>=0.19.0",
    "psutil>=5.8.0",
    "pytest>=7.0.0",
]
keywords = ["telegram", "bot", "faq", "semantic-search", "ai", "nlp"]

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "mypy",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov",
    "pytest-asyncio",
]

[project.urls]
Homepage = "https://github.com/your-username/faq-bot"
Documentation = "https://github.com/your-username/faq-bot/blob/main/README.md"
Repository = "https://github.com/your-username/faq-bot.git"
"Bug Tracker" = "https://github.com/your-username/faq-bot/issues"

[project.scripts]
faq-bot = "src.main:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
namespaces = false

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt", "*.json", "*.db"]